+++ src/configurationConstants.ts
@@ -17,3 +17,5 @@ export const TREEVIEW_PAGINATION = "treeView.pagination";
 export const FORMS_EXPORT_FILENAME_CONF = "export.forms.filename";
 export const FORM_EXPORT_FILENAME_CONF = "export.form.filename";
 
+export const WORKFLOW_EXPORT_FILENAME_CONF = "export.workflow.filename";
+